pyproject.toml
[project]
name = "many-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "phoenix-agents>=0.1.4",
]
